Add mid-q score and projection for warpfork
(I guess projections should've been earlier?  oh well)

The selector grammar OKRs are a tad interesting and I wonder if we should rephrase those, because we're actually pursuing a specific implementation plan that sliiiightly eschews any custom grammar efforts for now... but that thing, it's going.

Calling IPLD Schema Experiment indeed demo'd :D  Much more fun to do there in the future, but we're actually using parts of those demos to iterate on drafting schemas for other things, so, they're already flat out providing value woohoo

go-ipld-prime is both still far from complete and yet probably ahead of where I would've expected it to be at the start of the year, so, that's not a typo either.