Add q2 scorings. (#13)
Somewhat arbitrarily, frankly.  These OKRs do not represent the
majority of the work done in the quarter, nor the goals we've been
operating on for the majority of the quarter (if, honestly, ever).
And the quarter would've been worse if we'd stuck any closer to these.

Side note, I'm not actually sure how we got to "go-ipld-prime is being
used by {filecoin,ipfs}".  I don't think I signed off on those.  So
those are flat zeros.  My last checkpoint in my personal notes says
"go-ipld-prime is in use by at least one major project" -- because what
we want is exercise, not necessarily to take on the biggest most-used
limelight integrations first! -- and that might've at least gotten a
partial completion score.

Unixfsv2 also shouldn't have said "usable".  I don't think there's any
way we would've expected "usable" to come out of this quarter.  More
drafts to get the ball rolling?  Sure.  (And those we have -- drafts
of schemas have made it possible to begin advancing that conversation;
and we've successfully united on the important concept that advanced
data layouts are a prerequisite for a the Unixfsv2 that we *want*
(e.g. that won't result in a subsequent "v3" shortly after).  *This*
is what I would've actually expected at the outset of the quarter; and
that's roughly what we got.)  Implementation?  Naw.  And realistically,
we'd probably rather make the go implementation come after
schema-powered codegen, because that's going to be a sizable leverage
multiplier, etc.

"go-ipld-prime can de-serialize common data formats into consistent
structs" was a shout out to either schema-powered codegen or to "bind"
node implementations using reflection.  The score on those is low
primarily because we continued to spend a lot of time this quarter on
selectors spec followthrough -- I don't even wanna touch why we were
too optimistic to include that in the OKRs as a real task -- as well
as spend a fair amount of time trying to figure out what advanced data
layouts are or are not going to encompass -- again, add that to the
rack of things not accounte for here.

Another thing that's missing from the OKR table but existed in my last
checkpoint of OKR drafts was "solid docs on the IPLD Data Model (and
what it's not) landed in ipld main repo" -- and I think that one is
both incredibly important, and we've made good progress.  I spent a lot
of time on communications this quarter, including recorded (thank you,
c3voc) conference talks, and this coverage has been very useful.
I'm frustrated it's not accounted for in OKR scoring; it was important.

Overall, we *still* need to get radically, radically better at making
conservative, reasonably-sized, incremental progress goals.  I feel
like we say this every quarter, And Yet Somehow...

Maybe OKRs aren't a process that leads to good things or causes
accurate communication.  Maybe quarter-sized fixed tasks are a bad
idea.  Maybe fixed-in-advanced tasks don't actually jive with work
that involves discovery processes.  Maybe a reporting structure that
sidelines/ignores anything not planned in advance is inherently going
to mis-report things -- and maybe that makes such a structure
fundamentally unfit at its own primary purpose.  Maybe.
Just... sayin'.