Added Mid-q scores for 2020/Q1
I continue to find OKRs a little wild as a format and questionable in
how well they apply to our scale of coordination, for the record.
There's no indications in this table of how big tasks are relative
to each other; no indications of dependencies; no indications of...
anything I'd care about in planning work or estimating its odds of
completion, honestly.  But hey.  Just me.