OKR score.
Interface and performance improvements landed, 100%.

Brought them to selectors: yes.
Is propagated to graphsync and downstreams: less yes.

(Possible lesson for future: might not be wise to target an OKR to
more than one repo at a time: odds of getting halted in one and
not the other because of transitive dependencies or other logistical
issues is higher than not.  At which point it becomes debatable how
to score progress, and how someone at a distance could interpret
that score.)